% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_data.R
\name{base_qtl_data}
\alias{base_qtl_data}
\title{Get base qtl data}
\usage{
base_qtl_data(obj)
}
\arguments{
\item{obj}{An object of class summary.wgaim or data.frame}
}
\value{
A data.frame
}
\description{
Get base qtl data
}

