\name{beadplot-package}
\alias{beadplot-package}
\alias{beadplot}
\docType{package}
\title{
\packageTitle{beadplot}
}
\description{
\packageDescription{beadplot}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{beadplot}
\packageIndices{beadplot}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{beadplot}

Maintainer: \packageMaintainer{beadplot}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
