% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_plots.R
\name{combo_plot}
\alias{combo_plot}
\title{Make main qtl plot}
\usage{
combo_plot(obj, qtl_summary, main_bead_size, main_bead_col, founder_bead_size,
  founder_bead_col)
}
\arguments{
\item{obj}{An object of class 'cross', 'map', 'mpcross', 'tidy_gen_map'}

\item{qtl_summary}{An object of class mpwgaim.summary or a data.frame}

\item{main_bead_size}{Character - name of the variable to map bead size to (default = qtl_perc_var)}

\item{main_bead_col}{Character - Name of the variable to map bead colour to (default = qtl_logp)}

\item{founder_bead_size}{Character - name of the variable mapped to bead size}

\item{founder_bead_col}{Character - name of the variable mapped to bead colour}
}
\value{
a ggplot object
}
\description{
Make main qtl plot
}

