% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{deFactorise}
\alias{deFactorise}
\title{Identifies factor variables in a data.frame and converts them to character}
\usage{
deFactorise(data, ignore.var = NULL)
}
\description{
@param data A data.frame.
 @param ignore.var Any variables to ignore. Default is NULL.
 @return The updated data.frame.
}

