% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfounders.R
\name{founder_names}
\alias{founder_names}
\title{Get founder names}
\usage{
founder_names(obj)
}
\arguments{
\item{obj}{An mpcross or mpInterval object}
}
\value{
The founder names (character vector)
}
\description{
Get founder names
}

