% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_data.R
\name{founder_qtl_data}
\alias{founder_qtl_data}
\title{Get founder qtl data}
\usage{
founder_qtl_data(obj)
}
\arguments{
\item{obj}{An object of class summary.wgaim or data.frame}
}
\value{
A data.frame
}
\description{
Get founder qtl data
}

