% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_coords.R
\name{get_lg_offsets}
\alias{get_lg_offsets}
\title{Get offsets for linkage groups}
\usage{
get_lg_offsets(map, order = NULL)
}
\arguments{
\item{map}{A map object}

\item{order}{A vector of linkage group names to sort map by}
}
\value{
offsets as a vector
}
\description{
Get offsets for linkage groups
}
