% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_plots.R
\name{get_long_coords}
\alias{get_long_coords}
\title{Get long version of map_coords}
\usage{
get_long_coords(obj)
}
\arguments{
\item{obj}{An object of class 'cross', 'map', 'mpcross', 'tidy_gen_map'}
}
\value{
A data.frame
}
\description{
Get long version of map_coords
}

