% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m4_cross_qtl}
\alias{m4_cross_qtl}
\title{A 4way mpcross object with qtl}
\format{A list of class 'mpcross' produced by the qtl package
\describe{
\item{founders}{A list of the genotyping data of the founders}
\item{finals}{A list of the genotyping data of the finals}
...
}}
\usage{
m4_cross_qtl
}
\description{
A cross produced by mpMap, with a qtl model
}
\keyword{datasets}

