% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m4_summary}
\alias{m4_summary}
\title{A summary.mpwgaim object}
\format{A list of class 'summary.mpwgaim' produced by the mpwgaim package
\describe{
\item{summary}{A data.frame of qtl output}
\item{LOGPtrace}{A list describing the LOGP trace of qtl detection}
...
}}
\usage{
m4_summary
}
\description{
Summary of qtl output from mpwgaim of m4_int_qtl
}
\keyword{datasets}
