% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_data.R
\name{main_qtl_data}
\alias{main_qtl_data}
\title{Get main effect qtl data}
\usage{
main_qtl_data(obj)
}
\arguments{
\item{obj}{An object of class summary.wgaim or data.frame}
}
\value{
A data.frame
}
\description{
Get main effect qtl data
}

