% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beadplot.R
\name{make_base_plot}
\alias{make_base_plot}
\title{Makes a linear ggplot object based on genetic map}
\usage{
make_base_plot(obj)
}
\arguments{
\item{obj}{An object of class 'cross', 'map', 'mpcross', 'tidy_gen_map'}
}
\value{
A ggplot object
}
\description{
Makes a linear ggplot object based on genetic map
}

