% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundation_plots.R
\name{make_founder_plot}
\alias{make_founder_plot}
\title{Make a founder plot}
\usage{
make_founder_plot(obj, yvar = NULL)
}
\arguments{
\item{obj}{An object of class 'cross', 'map', 'mpcross', 'tidy_gen_map'}

\item{yvar}{Variable to map y to (character)}
}
\value{
A ggplot object
}
\description{
Make a founder plot
}
\examples{
make_founder_plot(m4_cross_qtl)
}

