% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_plots.R
\name{make_main_plot}
\alias{make_main_plot}
\title{Makes a linear ggplot object based on genetic map}
\usage{
make_main_plot(obj)
}
\arguments{
\item{obj}{An object of class 'cross', 'map', 'mpcross', 'tidy_gen_map'}
}
\value{
A ggplot object
}
\description{
Makes a linear ggplot object based on genetic map
}

