% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_plots.R
\name{make_main_qtl_plot}
\alias{make_main_qtl_plot}
\title{Make main qtl plot}
\usage{
make_main_qtl_plot(obj, qtl_summary, main_bead_size = "qtl_perc_var",
  main_bead_col = "qtl_logp")
}
\arguments{
\item{obj}{An object of class 'cross', 'map', 'mpcross', 'tidy_gen_map'}

\item{qtl_summary}{An object of class mpwgaim.summary or a data.frame}

\item{main_bead_size}{Character - name of the variable to map bead size to (default = qtl_perc_var)}

\item{main_bead_col}{Character - Name of the variable to map bead colour to (default = qtl_logp)}
}
\value{
a ggplot object
}
\description{
Make main qtl plot
}

