% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_data.R
\name{mpwgaim_summary_tidy}
\alias{mpwgaim_summary_tidy}
\title{Tidy an mpwgaim summary}
\usage{
mpwgaim_summary_tidy(obj)
}
\arguments{
\item{obj}{An object of class mpwgaim.summary or a data.frame}
}
\value{
A tibble
}
\description{
Tidy an mpwgaim summary
}

