% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{na.locf}
\alias{na.locf}
\title{A stripped down version of zoo::na.locf()}
\usage{
na.locf(object)
}
\arguments{
\item{object}{a vector}
}
\value{
vector with NA removed
}
\description{
A stripped down version of zoo::na.locf()
}

