% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfounders.R
\name{nfounders}
\alias{nfounders}
\title{Get number of founders of a multi-parental object}
\usage{
nfounders(obj)
}
\arguments{
\item{obj}{A map or cross object}
}
\value{
The number of founders
}
\description{
Get number of founders of a multi-parental object
}

