% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{nonNumeric}
\alias{nonNumeric}
\title{Checks if any non-numeric elements are present}
\usage{
nonNumeric(x)
}
\arguments{
\item{x}{a vector to be inspected}
}
\value{
TRUE or FALSE
}
\description{
Checks if any non-numeric elements are present
}
\examples{
vec1 <- 1:10
vec2 <- c(vec1, "A")
nonNumeric(vec1)
nonNumeric(vec2)
}

