% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_mpcross}
\alias{small_mpcross}
\title{A small 4way mpcross object with qtl}
\format{A list of class 'mpcross' produced by the qtl package
\describe{
\item{founders}{A list of the genotyping data of the founders}
\item{finals}{A list of the genotyping data of the finals}
...
}}
\usage{
small_mpcross
}
\description{
A cross produced by mpMap, with a qtl model
}
\keyword{datasets}

