% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_mpint}
\alias{small_mpint}
\title{A small 4way mpInterval object}
\format{A list of class 'mpInterval' produced by the mpwgaim package
\describe{
\item{pheno}{A phenotype data.frame}
\item{geno}{A list of the genotyping data}
...
}}
\usage{
small_mpint
}
\description{
A cross produced by mpwgaim, with a qtl model
}
\keyword{datasets}

