% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{autoNumeric}
\alias{autoNumeric}
\title{Converts character data.frame variables to numeric if it will not coerce to NA}
\usage{
autoNumeric(data, ignore.var = NULL)
}
\arguments{
\item{data}{A data.frame.}

\item{ignore.var}{Any variables to ignore. Default is NULL.}
}
\value{
The updated data.frame.
}
\description{
Converts character data.frame variables to numeric if it will not coerce to NA
}
