% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{deFactorise}
\alias{deFactorise}
\title{Identifies factor variables in a data.frame and converts them to character}
\usage{
deFactorise(data, ignore.var = NULL)
}
\arguments{
\item{data}{a data.frame}

\item{ignore.var}{any variables to ignore. Default is NULL}
}
\value{
The updated data.frame
}
\description{
Identifies factor variables in a data.frame and converts them to character
}

