% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{deFactorise}
\alias{deFactorise}
\title{Identifies factor variables in a data.frame and converts them to character}
\usage{
deFactorise(data, ignore.var = NULL)
}
\arguments{
\item{data}{A data.frame.}

\item{ignore.var}{Any variables to ignore. Default is NULL.}
}
\value{
The updated data.frame.
}
\description{
Identifies factor variables in a data.frame and converts them to character
}
