% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_plots.R
\name{make_main_qtl_plot}
\alias{make_main_qtl_plot}
\title{Make main qtl plot}
\usage{
make_main_qtl_plot(obj, qtl_summary)
}
\arguments{
\item{obj}{An object of class 'cross', 'map', 'mpcross', 'tidy_gen_map'}

\item{qtl_summary}{An object of class mpwgaim.summary or a data.frame}
}
\value{
a ggplot object
}
\description{
Make main qtl plot
}

